pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "kedro-openlineage"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
description = "OpenLineage integration for Kedro"
dependencies = [
    "kedro>=0.19.11",
    "openlineage-python>=1.28.0",
    "structlog>=25.1.0",
]
authors = [
    {name = "Juan Luis Cano Rodríguez", email = "juan_luis_cano@mckinsey.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dynamic = ["version"]

[project.urls]
source = "https://github.com/astrojuanlu/kedro-openlineage"
tracker = "https://github.com/astrojuanlu/kedro-openlineage/issues"
documentation = "https://kedro-openlineage.readthedocs.io"

[project.optional-dependencies]
test = [
    "mypy",
    "hypothesis",
    "pytest",
    "pytest-cov",
]
doc = [
    "furo",
    "myst-parser",
    "sphinx>=5",
    "sphinx-copybutton",
]

[project.entry-points."kedro.hooks"]
plugin_name = "plugin_name.plugin:hooks"

[tool.pdm.version]
source = "scm"

[tool.ruff]
show-fixes = true

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # Pycodestyle
    "W",  # Pycodestyle
    # "D",  # Pydocstyle
    "UP",  # pyupgrade
    "I",  # isort
    "PL",  # Pylint
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
python_version = "3.8"
warn_redundant_casts = true
warn_unused_configs = true
pretty = true
show_error_codes = true
show_error_context = true
show_column_numbers = true

disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true

# More strict checks for library code
[[tool.mypy.overrides]]
module = "kedro_openlineage"
disallow_untyped_defs = true

# Ignore certain missing imports
# [[tool.mypy.overrides]]
# module = "thirdparty.*"
# ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
